fix: temporary do not support reboot on pearl

+++ orb-jobs-agent/src/handlers/reboot.rs
@@ -1,8 +1,9 @@
-use crate::job_system::ctx::Ctx;
+use crate::job_system::ctx::{Ctx, JobExecutionUpdateExt};
 use color_eyre::{
-    eyre::{bail, eyre},
+    eyre::{bail, eyre, Context as _},
     Result,
 };
+use orb_info::orb_os_release::OrbOsPlatform;
 use orb_relay_messages::jobs::v1::{JobExecutionStatus, JobExecutionUpdate};
 use std::{future::Future, path::Path};
 use tokio::{fs, io};
@@ -11,6 +12,19 @@ use tracing::info;
 /// command format: `reboot`
 #[tracing::instrument]
 pub async fn handler(ctx: Ctx) -> Result<JobExecutionUpdate> {
+    let os_release_path = &ctx.deps().settings.os_release_path;
+    let os_release_contents = fs::read_to_string(os_release_path)
+        .await
+        .context("failed to read Orb OS release file")?;
+    let os_release = orb_info::orb_os_release::OrbOsRelease::parse(os_release_contents)
+        .context("failed to parse Orb OS release information")?;
+
+    if os_release.orb_os_platform_type == OrbOsPlatform::Pearl {
+        return Ok(ctx
+            .status(JobExecutionStatus::FailedUnsupported)
+            .stderr("reboot is not supported on Pearl devices"));
+    }
+
     run_reboot_flow(ctx, "reboot", |_ctx| async move {
         Ok(RebootPlan::with_stdout("rebooting\n"))
     })
